Extract DOM mutation event processing into utility class
